Started work on downloadable Archive with shared components. Split Insights into a wrapper component (for main client app) and body (shared with the Archive). Added new webpack config for building archive via package.json script.
