+++ src/client/components/Archive/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="referrer" content="no-referrer">
     <link rel="apple-touch-startup-image" href="/userData/images/dn.png" />
-    <title>DocNow</title>
+    <title>DocNow - Archived Collection</title>
     <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,700" rel="stylesheet">
     <script type="module" src="https://unpkg.com/ionicons@4.5.10-0/dist/ionicons/ionicons.esm.js"></script>
     <script nomodule="" src="https://unpkg.com/ionicons@4.5.10-0/dist/ionicons/ionicons.js"></script>
@@ -18,5 +18,5 @@
 </head>
 <body>
     <div id="app"></div>
-<script type="text/javascript" src="/main-bd0364e61370860071ae.js"></script></body>
+</body>
 </html>
