+++ src/client/components/Archive/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Archive from './Archive';
+
+ReactDOM.render(<Archive />, document.getElementById('app'));
